clean up dependencies
