+++ oceanics-io-fcns/src/sensor-things.ts
@@ -2,17 +2,28 @@
  * Cloud function version of API
  */
 import type { Handler } from "@netlify/functions";
-import { catchAll, connect, tokenClaim, parseFunctionsPath, getLabelIndex, fetchLinked } from "./shared/driver";
+import { catchAll, connect, tokenClaim, parseFunctionsPath, transform } from "./shared/driver";
 import { Node, Links } from "./shared/pkg/neritics";
 
 /**
  * Get an array of all collections by Node type
  */
 const index = async () => {
+
+  const { query } = Node.allLabels();
+  const { records } = await connect(query);
+  const restricted = new Set(["Provider", "User"]);
+  //@ts-ignore
+  const fields = new Set(records.flatMap(({ _fields: [label] }) => label).filter(label => !restricted.has(label)));
+  const data = [...fields].map((label: string) => Object({
+    name: label,
+    url: `/api/${label}`
+  }));
+
   return {
     statusCode: 200,
     headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(getLabelIndex())
+    body: JSON.stringify(data)
   };
 }
 
@@ -38,7 +49,8 @@ const create = async (left: Node, right: Node) => {
  * by any single property. 
  */
 const metadata = async (left: Node, right: Node) => {
-  const value = await fetchLinked(left, right);
+  const { query } = (new Links()).query(left, right, right.symbol);
+  const value = transform((await connect(query))).map(node => node[1]);
   return {
     statusCode: 200,
     headers: { "Content-Type": "application/json" },
@@ -72,7 +84,6 @@ const mutate = (left: Node, right: Node) => {
  * 
  */
 const remove = async (left: Node, right: Node) => {
-
   const link = new Links();
   const { query } = link.deleteChild(left, right);
   await connect(query)
@@ -81,32 +92,12 @@ const remove = async (left: Node, right: Node) => {
   }
 }
 
-const join = (left: Node, right: Node) => {
+const join = async (left: Node, right: Node) => {
   return {
-    statusCode: 204
+    statusCode: 501,
+    body: JSON.stringify({ message: "Not Implemented" })
   }
 }
-
-
-
-//     # Generate the Cypher query
-//     # pylint: disable=eval-used
-//     cypher = Links(
-//         label="Join",
-//         **body
-//     ).join(
-//         *parse_as_nodes((
-//             eval(root)(uuid=rootId),
-//             eval(entity)(uuid=uuid)
-//         ))
-//     )
-
-//     # Execute transaction and end session before reporting success
-//     with db.session() as session:
-//         session.write_transaction(lambda tx: tx.run(cypher.query))
-
-//     return None, 204
-
 
 const drop = async (left: Node, right: Node) => {
   const cypher = (new Links()).drop(left, right)
@@ -120,19 +111,12 @@ const drop = async (left: Node, right: Node) => {
  * Retrieve nodes that are linked with the left entity
  */
 const topology = (left: Node, right: Node) => {
-  // const link = new Link()
-  // const cypher = link.query()
-
-  //     nodes = ({"cls": root, "id": rootId}, {"cls": entity})
-
-  //     # Pre-calculate the Cypher query
-  //     cypher = Links().query(*parse_as_nodes(nodes), "b")
-
-  //     with db.session() as session:
-  //         value = [*map(lambda x: x.serialize(), session.write_transaction(lambda tx: tx.run(cypher.query)))]
-
-  //     return {"@iot.count": len(value), "value": value}, 200
+  return {
+    statusCode: 501,
+    body: JSON.stringify({ message: "Not Implemented" })
+  }
 }
+
 
 /**
  * Browse saved results for a single model configuration.
