+++ oceanics-io-fcns/src/shared/data.ts
@@ -4,6 +4,7 @@ import { createHash } from "crypto";
 // import NetCDFReader from 'netcdfjs';
 import { readFileSync, writeFileSync } from 'fs';
 
+
 interface IQuery {
     service: string;
     key: string;
